Moved App to components folder
