+++ frontend/index.js
@@ -1,4 +1,4 @@
 import { registerRootComponent } from 'expo';
-import App from './App';
+import App from './components/App';
 
 registerRootComponent(App);
